\name{getExpDescription}
\alias{getExpDescription}

\title{ getExpDescription
}
\description{
Gets the \code{bio\_sample} table from the database. May be used as \code{phenoData}.
}
\usage{
getExpDescription()
}
\value{ Table of experimental factors assigned to numbers of samples. 
}
\author{
   Michal Okoniewski, Anna Lesniewska
}
