\name{getSumsExp}
\alias{getSumsExp}

\title{ getSumsExp
}
\description{
Gets the sum of reads in all the samples present in the database in the \code{seq\_read} table
}
\usage{
getSumsExp()
}
\value{ Vector of sums
}
\author{
   Michal Okoniewski, Anna Lesniewska
}
\examples{
#  if (xmapConnected())  
#  {
#   sums <- getSumsExp()
#   nsums
#  }
}
