\name{rs.list}
\alias{rs.list}
\alias{sample_data_rnaSeqMap}
\docType{data}
\title{Example of sequencing data for rnaSeqMap library}
\description{ A fragment of sequencing data from 6 samples - human. 
}
\usage{data(sample_data_rnaSeqMap)}
\format{
 A list with 17 SeqReads objects,each with sequencing reads from 6 samples
 sequenced with ABI SOLID machine.
}
\examples{
#  data(sample_data_rnaSeqMap)
#  length(rs.list)
#  gene1rs <- rs.list[[1]]
}
\keyword{datasets}
