\name{simplePlot}
\Rdversion{1.1}
\alias{simplePlot}
\title{simplePlot - quick plot for the coverages  
}
\description{ Plots 2 or 3 coverages with fixed colors.  


}
\usage{
simplePlot (nd, exps, xlab="genome coordinates", ylab="coverage")
}
\arguments{
  \item{nd}{ NucleotideDistr object to plot 
}
  \item{exps}{ Samples to plot - numeric vector
}
  \item{xlab}{    
}
  \item{ylab}{   
}
}

\author{
   Michal Okoniewski
}
